charts use metrics
